+++ lib/src/certificates_handler_io.dart
@@ -77,16 +77,27 @@ class CertificatesHandlerIO extends CertificatesHandler {
   String _pathFileName(String path) => pack_path.split(path).last;
 
   @override
-  Future<SecurityContext?> buildSecurityContext(List<Domain> domains,
-      {bool loadAllHandledDomains = true}) async {
+  Future<Map<String, SecurityContext>?> buildSecurityContexts(
+      List<Domain> domains,
+      {bool allowUnresolvedDomain = false,
+      bool loadAllHandledDomains = true}) async {
     final securityContextBuilder = SecurityContextBuilder();
+
+    var resolvedDomains = <Domain>{};
+    var unresolvedDomains = <Domain>{};
 
     for (final domain in domains) {
       final domainOk =
           await _useDomainCertificate(securityContextBuilder, domain.name);
-      if (!domainOk) {
-        return null;
+      if (domainOk) {
+        resolvedDomains.add(domain);
+      } else {
+        unresolvedDomains.add(domain);
       }
+    }
+
+    if (unresolvedDomains.isNotEmpty && !allowUnresolvedDomain) {
+      return null;
     }
 
     if (loadAllHandledDomains) {
@@ -99,8 +110,8 @@ class CertificatesHandlerIO extends CertificatesHandler {
       }
     }
 
-    final securityContext = securityContextBuilder.build();
-    return securityContext;
+    final securityContexts = securityContextBuilder.buildAll();
+    return securityContexts;
   }
 
   Future<bool> _useDomainCertificate(