+++ lib/src/letsencrypt.dart
@@ -266,6 +266,8 @@ class LetsEncrypt {
   }
 
   @Deprecated('Use `startServer`. Will be removed at v2.1.0')
+
+  /// Use [startServer].
   Future<List<HttpServer>> startSecureServer(
       Handler handler, Map<String, String> domainsAndEmails,
       {int? backlog,
@@ -294,15 +296,18 @@ class LetsEncrypt {
   /// Starts 2 [HttpServer] instances, one HTTP at [port]
   /// and other HTTPS at [securePort].
   ///
-  /// - If [checkCertificate] is `true`, will check the current certificate
-  /// - if [requestCertificate] is `true` then we will  acquire/renew the certificate
+  /// - If [checkCertificate] is `true`, will check the current certificates.
+  /// - if [requestCertificate] is `true` then we will  acquire/renew the certificates
   ///   as needed.
-  /// - If [forceRequestCertificate] is `true` then we will force the acquistion
-  ///   of a new certificate.
+  /// - If [forceRequestCertificate] is `true` then we will force the acquisition
+  ///   of a new certificates.
   ///
-  /// *WARNING: the Lets Encrypt CA has VERY tight rate limits
-  /// on certificate acquistion. If you breach them you will not be able to
-  /// acquire a new production certificate for 168 hours!!!*
+  /// *WARNINGS:*
+  /// - *The Lets Encrypt CA has VERY tight rate limits
+  ///   on certificate acquisition. If you breach them you will not be able to
+  ///   acquire a new production certificate for 168 hours!!!*
+  /// - *Only use `requestCertificate: true` or `forceRequestCertificate: true`
+  ///   if you are certain that you won't make unnecessary certificate requests.*
   Future<({HttpServer http, HttpServer https})> startServer(
       Handler handler, List<Domain> domains,
       {int? backlog,
@@ -331,9 +336,19 @@ class LetsEncrypt {
     final server = await serve(handlerWithChallenge, bindingAddress, port,
         backlog: backlog, shared: shared);
 
-    Future<HttpServer> startSecureServer(SecurityContext securityContext) =>
-        serve(handlerWithChallenge, bindingAddress, securePort,
-            securityContext: securityContext, backlog: backlog, shared: shared);
+    Future<HttpServer> startSecureServer(
+        Map<String, SecurityContext> securityContexts,
+        {SecurityContext? defaultSecurityContext}) {
+      defaultSecurityContext ??= securityContexts['*'] ??
+          securityContexts.entries.firstOrNull?.value ??
+          (throw ArgumentError(
+              "Can't define `defaultSecurityContext`> null `defaultSecurityContext` and empty `securityContexts`"));
+
+      return serve(handlerWithChallenge, bindingAddress, securePort,
+          securityContext: defaultSecurityContext,
+          backlog: backlog,
+          shared: shared);
+    }
 
     HttpServer? secureServer;
 
@@ -341,22 +356,29 @@ class LetsEncrypt {
     logger.info(
         'Handled domains: ${certificatesHandler.listAllHandledDomains()}');
 
-    var securityContext = await certificatesHandler.buildSecurityContext(
+    var securityContexts = await certificatesHandler.buildSecurityContexts(
         domains,
+        allowUnresolvedDomain: false,
         loadAllHandledDomains: loadAllHandledDomains);
 
     logger.info(
-        '''securityContext[loadAllHandledDomains: $loadAllHandledDomains]: $securityContext''');
+        '''securityContext[loadAllHandledDomains: $loadAllHandledDomains]: $securityContexts''');
 
-    if (securityContext == null) {
+    if (securityContexts == null || securityContexts.isEmpty) {
       if (!requestCertificate) {
-        throw StateError(
-            """No previous SecureContext. Parameter `requestCertificate` is `false`, can't request certificate!""");
+        if (securityContexts == null) {
+          throw StateError(
+              """Can't load all `SecurityContext`s. Parameter `requestCertificate` is `false`, can't request certificates! Domains: ${Domain.toNames(domains)}""");
+        } else {
+          throw StateError(
+              """No previous `SecurityContext`s. Parameter `requestCertificate` is `false`, can't request certificates! Domains: ${Domain.toNames(domains)}""");
+        }
       }
 
       final domainsToCheck = certificatesHandler.listNotHandledDomains(domains);
 
-      logger.info('Requesting certificate for: $domainsToCheck');
+      logger.info(
+          'Requesting certificate for: ${Domain.toNames(domainsToCheck)}');
 
       for (final domain in domainsToCheck) {
         final ok = await this.requestCertificate(domain);
@@ -365,18 +387,20 @@ class LetsEncrypt {
         }
       }
 
-      securityContext = await certificatesHandler.buildSecurityContext(domains,
+      securityContexts = await certificatesHandler.buildSecurityContexts(
+          domains,
+          allowUnresolvedDomain: false,
           loadAllHandledDomains: loadAllHandledDomains);
-      if (securityContext == null) {
+      if (securityContexts == null || securityContexts.isEmpty) {
         throw StateError(
-            '''Error loading SecureContext after successful request of certificate for: $domains''');
+            '''Error loading SecureContext after successful request of certificates> domainsToCheck: ${Domain.toNames(domainsToCheck)} ; domains: ${Domain.toNames(domains)}''');
       }
 
       logger.info(
           'Starting secure server> port: $securePort ; domains: $domains');
-      secureServer = await startSecureServer(securityContext);
+      secureServer = await startSecureServer(securityContexts);
     } else {
-      secureServer = await startSecureServer(securityContext);
+      secureServer = await startSecureServer(securityContexts);
 
       if (checkCertificate) {
         logger.info('Checking domains certificates: $domains');
@@ -402,17 +426,17 @@ class LetsEncrypt {
 
         if (refreshedCertificate) {
           logger.warning('Refreshing SecureContext due new certificate.');
-          securityContext = await certificatesHandler.buildSecurityContext(
+          securityContexts = await certificatesHandler.buildSecurityContexts(
               domains,
               loadAllHandledDomains: loadAllHandledDomains);
-          if (securityContext == null) {
+          if (securityContexts == null || securityContexts.isEmpty) {
             throw StateError(
                 '''Error loading SecureContext after successful certificate check for: ${Domain.toNames(domains)}''');
           }
 
           logger.warning('Restarting secure server...');
           await secureServer.close(force: true);
-          secureServer = await startSecureServer(securityContext);
+          secureServer = await startSecureServer(securityContexts);
         }
       }
     }